Fix a bug with component declarations in entities
This bug caused the component declaration to be omitted from the list of
VHDL objects and caused the component ports to be added to the entity
port.

The root cause was the end of the entity declaration not being parsed
correctly when this was specified as `end entity xyz;` instead of `end xyz;`.

Another bug was that there was no token to identify component
declarations inside entity architectures.

Signed-off-by: Wouter van Verre <woutervanverre@gmail.com>